add JS Exercise - arrays-objects
